PennyBot V2.5
PennyBot now designed to run on the Lamda web services from Amazon.

Comment and Submission streaming have been removed, althrough legacy code is being left in seperate methods. Lamda does not support constant running, so a return to Iterative.

The Database functions and SQL backend have been offloaded to the Amazon Database system.

Additional packages and dependencies are needed for Pennybot to run in a lamda container, all PIP.

Old Command Structure has been removed and replaced with Json Lookup and editing of arrays. Allows for on the fly modification of commands and resources.

All elements and data are hosted in S3.